component name change
